requirements.txt
Django==2.0.5                   # Our web framework of choice
djangorestframework==3.8.2      # A library for implementing APIs within Django
django-extensions==2.0.7        # A utility library for enhancing Django development
